requirements.txt
py-clob-client>=0.9.0
websockets>=11.0
python-dotenv>=1.0.0
requests>=2.28.0
flask>=2.0.0
gunicorn>=20.0.0
